python/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
python-dotenv==1.0.0
pymongo==4.3.3
openai==0.27.4
pytz==2023.3